change falot to decimal of promotionals table

+++ db/migrate/20190211124306_create_promotionals.rb
@@ -1,8 +1,8 @@
 class CreatePromotionals < ActiveRecord::Migration[5.2]
   def change
     create_table :promotionals do |t|
-      t.float :cart_item_discount, default: 1
-      t.float :cart_discount, default: 1
+      t.decimal :cart_item_discount, default: 1
+      t.decimal :cart_discount, default: 1
       t.integer :item_p, default: 3
       t.integer :vendor_p, default: 3
       t.integer :order_discount, default: 1000
